book: Restructure lib
* mv `db::models::*` to `rpc::models::*`
* mv `fn rpc_server(...)` to `mod rpc`
* replace `pub mod ...` with `pub use crate::rpc::*`
